Refactor searchable and categorizable combobox to general method In case we want to use it for other parts of the UI :)
